% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRIF.R
\name{calculateEiA}
\alias{calculateEiA}
\title{calculates expression levels according to phenodata in transcriptomic data}
\usage{
calculateEiA(eSet, target.factor, A)
}
\arguments{
\item{eSet}{an ExpressionSet}

\item{target.factor}{a factor of interest}

\item{A}{a level of a factor (e.g., condition 1)}
}
\value{
matrix
}
\description{
This function calculates expression levels of each gene in one condition (A).
}
\examples{
data(ToniData)

formula <- geno1~geno2
samp <- list(A=as.character(formula[[2]]), B=as.character(formula[[3]]))
target.factor <- "Genotype"
EiA <- calculateEiA(eSet=ToniData, target.factor=target.factor, samp$A)

}
\author{
Kevin Rue-Albrecht \url{https://github.com/kevinrue/HudsonRIF}
cosmetical changes by Atsushi Fukushima
}
