% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRIF.R
\name{calculateEiAB}
\alias{calculateEiAB}
\title{calculates expression levels according to phenodata in expressionSet}
\usage{
calculateEiAB(eSet, target.factor, A, B)
}
\arguments{
\item{eSet}{an ExpressionSet}

\item{target.factor}{a factor of interest}

\item{A}{a level of a factor (e.g., condition 1)}

\item{B}{a level of a factor (e.g., condition 2)}
}
\value{
matrix
}
\description{
This function calculates expression levels of each gene 
in each data subset.
}
\examples{
data(ToniData)

formula <- geno1~geno2
samp <- list(A=as.character(formula[[2]]), B=as.character(formula[[3]]))
target.factor <- "Genotype"
EiAB <- calculateEiAB(eSet=ToniData, target.factor=target.factor, samp$A, samp$B)

}
\author{
Kevin Rue-Albrecht \url{https://github.com/kevinrue/HudsonRIF}
cosmetical changes by Atsushi Fukushima
}
