% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRIF.R
\name{calculateRIF1}
\alias{calculateRIF1}
\title{calculates RIF1}
\usage{
calculateRIF1(PIFi, dCij, DEGs)
}
\arguments{
\item{PIFi}{PIF}

\item{dCij}{a differential coexpression matrix}

\item{DEGs}{a list of DEGs (differentially expressed genes)}
}
\value{
vector
}
\description{
This function calculates RIF1.
}
\examples{
data(ToniData)

#formula <- geno1~geno2
#samp <- list(A=as.character(formula[[2]]), B=as.character(formula[[3]]))
#target.factor <- "Genotype"
#EiA <- calculateEiA(eSet=ToniData, target.factor=target.factor, samp$A)

}
\author{
Kevin Rue-Albrecht \url{https://github.com/kevinrue/HudsonRIF}
cosmetical changes by Atsushi Fukushima
}
