% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRIF.R
\name{calculateRIF2}
\alias{calculateRIF2}
\title{calculates RIF2}
\usage{
calculateRIF2(EiAB, rAij, rBij, DEGs, samp)
}
\arguments{
\item{EiAB}{an expression level}

\item{rAij}{a coexpression matrix in A}

\item{rBij}{a coexpression matrix in B}

\item{DEGs}{a list of DEGs (differentially expressed genes)}

\item{samp}{conditions}
}
\value{
vector
}
\description{
This function calculates RIF2.
}
\examples{
data(ToniData)

#formula <- geno1~geno2
#samp <- list(A=as.character(formula[[2]]), B=as.character(formula[[3]]))
#target.factor <- "Genotype"
#EiA <- calculateEiA(eSet=ToniData, target.factor=target.factor, samp$A)

}
\author{
Atsushi Fukushima
}
