% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{diffCoexPIF.plot}
\alias{diffCoexPIF.plot}
\title{Co-expression relationships between the specified TF and the DE}
\usage{
diffCoexPIF.plot(rRIF, reg.gene, pch.max = 0)
}
\arguments{
\item{rRIF}{"rRIF" class object}

\item{reg.gene}{The regulatory gene/probeset to plot co-expression values}

\item{pch.max}{A scaling factor defining the size of the data point}
}
\value{
none
}
\description{
Plots for the coexpression of the specified gene with each of the DE.
}
\details{
Modified Hudson::diffCoexPIF.plot()
(https://github.com/kevinrue/HudsonRIF)
}
\examples{
data(ToniData)
data(ToniData.DEGs)
data(ToniData.TFs)

res <- rRIF(eSet = ToniData,
            formula = geno1~geno2,
            target.factor = "Genotype",
            DEGs = ToniData.DEGs,
            cor.method="pearson",
            regulator.list = ToniData.TFs
            )
iMAplot(res)
}
