% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{iTable}
\alias{iTable}
\title{an interactive table of rRIF object}
\usage{
iTable(rRIF, top = 30)
}
\arguments{
\item{rRIF}{"rRIF" class object}

\item{top}{specifies top X RIF}
}
\description{
This function provides an interactive table of rRIF results
}
\examples{
data(ToniData)
data(ToniData.DEGs)
data(ToniData.TFs)

res <- rRIF(eSet = ToniData,
            formula = geno1~geno2,
            target.factor = "Genotype",
            DEGs = ToniData.DEGs,
            cor.method="pearson",
            regulator.list = ToniData.TFs
            )
iTable(res)
}
