% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRIF.R
\name{parameter.check}
\alias{parameter.check}
\title{validates all parameters}
\usage{
parameter.check(eSet, formula, target.factor, DEGs, cor.method, regulator.list)
}
\arguments{
\item{eSet}{an ExpressionSet}

\item{formula}{an object of class "formula" (e.g., genotype1~genotype2)}

\item{target.factor}{a factor of interest (e.g., "Genotype")}

\item{DEGs}{a list of DEGs (differentially expressed genes)}

\item{cor.method}{a list of correlation methods [i.e. c("pearson", "spearman", "kendall")]}

\item{regulator.list}{a list of transcriptional regulators (e.g., transcription factors)}
}
\value{
vector
}
\description{
This function calculates RIF2.
}
\examples{
data(ToniData)
data(ToniData.DEGs)
data(ToniData.TFs)

res <- parameter.check(
         eSet = ToniData,
         formula = geno1~geno2,
         target.factor = "Genotype",
         DEGs = ToniData.DEGs,
         cor.method="pearson",
         regulator.list = ToniData.TFs
         )
}
\author{
Kevin Rue-Albrecht \url{https://github.com/kevinrue/HudsonRIF}
cosmetical changes by Atsushi Fukushima
}
