% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRIF.R
\name{rRIF}
\alias{rRIF}
\title{calculates RIF1 and RIF2 in transcriptomic data}
\usage{
rRIF(eSet, formula, target.factor, DEGs, cor.method = "spearman",
  regulator.list = "")
}
\arguments{
\item{eSet}{an ExpressionSet}

\item{formula}{an object of class "formula" (e.g., genotype1~genotype2)}

\item{target.factor}{a factor of interest (e.g., "Genotype")}

\item{DEGs}{a list of DEGs (differentially expressed genes)}

\item{cor.method}{a list of correlation methods [i.e. c("pearson", "spearman", "kendall")]}

\item{regulator.list}{a list of transcriptional regulators (e.g., transcription factors)}
}
\value{
"rRIF" class object
}
\description{
This function calculates Regulatory Impact Factors (RIF1 and RIF2)
}
\examples{
data(ToniData)
data(ToniData.DEGs)
data(ToniData.TFs)

res <- rRIF(eSet = ToniData,
            formula = geno1~geno2,
            target.factor = "Genotype",
            DEGs = ToniData.DEGs,
            cor.method="pearson",
            regulator.list = ToniData.TFs
            )
}
\references{
Reverter A et al. Bioinformatics 26:896 (2010)
(\href{https://www.ncbi.nlm.nih.gov/pubmed/20144946}{PubMed})
}
\author{
Kevin Rue-Albrecht \url{https://github.com/kevinrue/HudsonRIF}
modified by Atsushi Fukushima
}
