% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToniData.DEGs}
\alias{ToniData.DEGs}
\title{A list of DEGs in chicken RNA-seq data.}
\description{
A list of differentially expressed genes (DEGs) based on Chicken RNA-seq 
generated by Bottje et al. (2017).
}
\details{
The number of DEGs here is 520.
}
\references{
Bottje et al. BMC Syst Biol 11:29 (2017)
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
