% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToniData}
\alias{ToniData}
\title{An example transcriptome dataset by Bottje et al. (2017)}
\description{
A transcriptome data matrix. The data is composed of 520 significant DEGs 
(differentially expressed genes) and 898 transcription factors generated 
by Bottje et al. (2017).
}
\details{
Twelve RNA-seq libraries were produced using breast muscle RNA samples 
from 6 HFE (high food efficiency) and 6 LFE (low FE) broilers. This 
chicken RNA-seq data were generated by mapping the reads to Gallus gallus
genome assembly version 4.
}
\references{
Bottje et al. BMC Syst Biol 11:29 (2017)
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
