% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToniData.TFs}
\alias{ToniData.TFs}
\title{A list of TFs in chicken RNA-seq data.}
\description{
A list of transcription factors (TFs) based on Chicken RNA-seq generated 
by Bottje et al. (2017).
}
\details{
The number of TFs here is 898.
}
\references{
Bottje et al. BMC Syst Biol 11:29 (2017)
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
